Rename mod to reflect its purpose
